+++ src/main/java/com/maxifly/vapi/model/REST/REST_photoAlbum.java
@@ -5,7 +5,7 @@ package com.maxifly.vapi.model.REST;
  */
 public class REST_photoAlbum {
 
-    public long id; // — идентификатор созданного альбома;
+    public int id; // — идентификатор созданного альбома;
     public int owner_id; // идентификатор пользователя или сообщества, которому принадлежит альбом;
     public String title; //— название альбома;
     public String description; // — описание альбома;
